sale(not done yet)

+++ DATA/DTOs/OrderItem/OrderItemDto.cs
@@ -6,7 +6,7 @@ namespace Deli.DATA.DTOs
         public Guid Id { get; set; }
         public int Quantity { get; set; }
         public string Name { get; set; }
-        public int? Price { get; set; }
+        public double? Price { get; set; }
         public string? Image { get; set; }
     }
 }
